support 1.3.0

+++ FreeSR.Proto/BlackInfo.cs
@@ -5,10 +5,10 @@ namespace FreeSR.Proto
 	[ProtoContract]
 	public class BlackInfo
 	{
-		[ProtoMember(1)] public long BeginTime;
-		[ProtoMember(2)] public long EndTime;
+		[ProtoMember(1)] public int BeginTime;
+		[ProtoMember(2)] public int EndTime;
 		[ProtoMember(3)] public int LimitLevel;
 		[ProtoMember(4)] public int BanType;
 	}
-	
+
 }
